fix: collision box too big error
