+++ common/src/main/kotlin/org/valkyrienskies/mod/common/VSGameUtils.kt
@@ -415,12 +415,14 @@ fun Level.getShipsIntersecting(aabb: AABB): Iterable<Ship> = getShipsIntersectin
 fun Level.getShipsIntersecting(aabb: AABBdc): Iterable<Ship> = allShips.getIntersecting(aabb, dimensionId)
 fun Level?.transformAabbToWorld(aabb: AABB): AABB = transformAabbToWorld(aabb.toJOML()).toMinecraft()
 fun Level?.transformAabbToWorld(aabb: AABBd) = this?.transformAabbToWorld(aabb, aabb) ?: aabb
-fun Level.transformAabbToWorld(aabb: AABBdc, dest: AABBd): AABBd {
+fun Level?.transformAabbToWorld(aabb: AABBdc, dest: AABBd): AABBd {
     val ship1 = getShipManagingPos(aabb.minX(), aabb.minY(), aabb.minZ())
+        ?: return dest.set(aabb)
     val ship2 = getShipManagingPos(aabb.maxX(), aabb.maxY(), aabb.maxZ())
+        ?: return dest.set(aabb)
 
     // if both endpoints of the aabb are in the same ship, do the transform
-    if (ship1 == ship2 && ship1 != null) {
+    if (ship1.id == ship2.id) {
         return aabb.transform(ship1.shipToWorld, dest)
     }
