+++ common/src/main/resources/valkyrienskies-common.mixins.json
@@ -125,7 +125,7 @@
     "client.gui.screens.MixinConnectScreen",
     "client.multiplayer.MixinClientPacketListener",
     "client.player.MixinLocalPlayer",
-    "client.player.MixinPlayer",
+    "feature.huge_bounding_box_fix.MixinPlayer",
     "client.renderer.MixinGameRenderer",
     "client.renderer.MixinLevelRenderer",
     "client.world.MixinClientChunkCache",
